add test demonstrating Loggers needs thread safety
Loggers.getLogger(nameOrClass) should return the same logger for the
same nameOrClass.  However, it is presently using an unprotected
HashMap, so this is not guaranteed, as demonstrated by the new test
added in this commit.

Beware that because of the nature of race conditions, it is not
guaranteed that this new LoggersTest.isThreadSafe() test will fail,
nor will it necessarily fail in the same way.  I have seen it pass; I
have seen it fail with a ConcurrentModificationException in
HashMap.computeIfAbsent(); I was expecting it to fail by returning
different loggers; and perhaps a very (very!) unlucky tester could see
the test never finish as two or more threads each happen to try to
resize the HashMap resulting in the linked list in a bucket becoming a
closed loop that both threads never stop going around.