fix thread safety of the Loggers class
Loggers.init() is a public method, so even though it is only used by
testing, there's no guarantee of that, and thus, to ensure that the
LOGGER_FACTORY field that it sets is memory safe, that field needs to be
volatile (or similar).

Moreover, ConsoleLoggerFactory should not be backed by a HashMap that is
not threadsafe, because Logger.getLogger() can be invoked by any thread.
With this change, I have not seen a LoggersTest.isThreadSafe() failure.